updated fix: https://issues.apache.org/jira/browse/AMQ-3982
Instead of locking on the PList iterator which forces a lock on every access to hasNext we should lock down at the PageFile level in readPage making the seek and read an atomic operation.  Since many reads on a single already loaded page don't need to be synchronized we pay a higher price for locking in the specific PList case than we do buy creating this atomic read operation in PageFile. 

git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@1375852 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 4bace216b91c9f8244976305f462fd212b0e20c3