fixes for: https://issues.apache.org/jira/browse/AMQ-4019
Refactoring the pools code to use commons-pool for more of the work and allowing us to
expose more stats info as well as enabling async connection pool cleanup of expired 
connections if so desired.  Added some additional tests.  We should continue to add some
more tests and perhaps add some more functionality to the Session pool in ConnectionPool
to let the commons-pool clean up Idle sessions after a time or to maintain the certain 
number of idle sessions etc.

git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@1383746 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: a59c9baaf546b13a9355e858d9043428c8556ad0