https://issues.apache.org/jira/browse/AMQ-4237
To support the bean renaming effort this pulls together the various code that create all our MBean names and puts them into one location which could later add supporting methods to create MBean queries for the unit tests etc.  This should make it simpler to iron out the naming as we go.  

git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@1430908 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 63a660ac04affba1d6f48da21b5de80b63a31a0a