https://issues.apache.org/jira/browse/AMQ-5290
Adds a subscription strategy model where the default is the normal
durable topic subscription based approach or a strategy that maps all
subscriptions and publish operations to a Virtual Topic model.  A
network of brokers can network the Queues instead of having the durable
topics subscriptions repaeted on each Broker.  

Former-commit-id: 413e4840d64fa8377103ce431a2a2c6cef6e0c20