https://issues.apache.org/jira/browse/AMQ-4814
Remove of a durable subscription when the keepDurableSubActive flag is
true (default) did not actually stop and cleanup the still active cursor
due to the fact that the cursor checks the flag of its subscription to
see if it should stay active when its stop method is called.  The
subscription needs to disable that flag when the subscription is being
removed so that the cursor will correctly shutdown. 

Former-commit-id: 7a68ad5d987caa20e6517beb12ee67d538b2ee08