changed permissions of "env" file
The previous permission of the  <activemq-install-dir>/bin/env file
(755) caused some problems on systems where the
<activemq-install-dir>/bin/ directory is part of the PATH variable.
The executable "env" is also available of all common unix platforms.
If <activemq-install-dir>/bin/ has precedence in the PATH environment
variable this causes lots of unexpected problems.
In future it would be a good idea to relocate the file to the
<activemq-install-dir>/conf directory.


Former-commit-id: 167c510331692bfa2648724d16330d49ec8dd0ae