https://issues.apache.org/jira/browse/AMQ-5712
Switching addMessageLast to tryAddMessageLast when messages are added
to a Queue pending cursor to allow a potential deadlock to be
avoided. There is more work to be done here but this will at least
prevent a deadlock from occurring.

Fix and test based off of a patch created by Timothy Bish.


Former-commit-id: cc6213ebf25a129b278a2ff0d7c32c25edd71eaa