https://issues.apache.org/jira/browse/AMQ-6088
The runtime plugins will now find the exact policy to update which means
that a destination can match more than one policy and the policy can
still be updated at runtime.

The java runtime broker also supports the ability to replace or add a
policy entry based on a flag on a new method call.


Former-commit-id: 9e7fae0d83c584f98e99024ba6d20e53f14b81f7