https://issues.apache.org/jira/browse/AMQ-6097
Fixing JournalCorruptionEofIndexRecoveryTest class

Because of the size being stored in the message order index, the test
testRecoveryAftercorruptionCheckSumExistingIndex was no longer passing
as that test was corrupting the size value in the journal, but not the
size value in the index.  The test now corrupts both values so that the
expect exception occurs again.


Former-commit-id: 507d40a79f15449ba6d79ef85d7fdf6bf50a3a65