Allow the AMQP test client to also be configure to trace frames
The test client can allow for quick tracing of the frame traffic via a
call to setTraceFrames on the client or connection instance before
connection to the remote.  This allows for tests to easily switch on /
off tracing.  The log4j.properties is also updated to output frame
tracing with the URI option is put on the AMQP transport or the client
value is enabled.  

Former-commit-id: 84cd815500bb486d2ab3d8724a76f0ac43dc3bf4