AMQ-6698 - Allow timeout of SSL handshake for auto+nio+ssl
The handshake now takes place inside of the async task that timesout so
that if the SSL handshake takes too long the connection attempt will
honor protocolDetectionTimeout and actually timeout.  Also, change the
default of protocolDetectionTimeout to 30 seconds so it matches the rest
of the broker's default connection timeout settings.


Former-commit-id: 2a0b785a041f35aa668d2371c9374eb0fd02063c