AMQ-6706 Allow for hooks before broker stop occurs
When Camel is run embedded within ActiveMQ it's shutdown is called at
the point the broker is already stopped, this can lead to data loss as
there could be exchanges that have not ended.
This adds the ability to specify preStopHook(s) that can be invoked
during ActiveMQ shutdown but before the broker is shutdown.

With this Camel can gracefully shutdown while the broker is still
started.


Former-commit-id: 155461a0ba0a082f5a8459f17ae6166bae950231