AMQ-7001 Resolve issues with encode failures on copied messages
Ensure that messages are copied to avoid contention on message content
when concurrent store and dispatch is enabled and better handle the AMQP
message type value.  Adds an AMQP Encoder for UTF8Buffer to encode down
to AMQP String encodings to allow for encoded OpenWire messages such as
MapMessage which can contain UTF8Buffer instances for String keys and
values.

Former-commit-id: 9ec6ee43b146736fdf635090b2829a0c621a00f3