AMQ-6949: SocketTimeoutException when using HTTP transport connector
A default socket timeout of 30 secods for the HTTP client was introduced with the following ticket: https://issues.apache.org/jira/browse/AMQ-6397

This timeout somewhow seems to conflict with the keep alive requests that stays open for arround 30 seconds. I increased the default timeout to 90 seconds and this seems to resolve the problem.

Former-commit-id: ea0bcc12f77594228bb270f199ba54b5dc3ce453